+++ interface/cli.js
@@ -1,5 +1,5 @@
 'use strict'
 
 var calc=require('../src/calculator.js');
-var res=calc.multiply(10,20);
+var res=calc.process('10*20');
 console.log(res);